Delay showing progress indicator in file chooser
For most projects the progress indicator would show for less than a tenth of a second, which would make it appear as flicker.

Also increase the initial poll interval to 0.02 seconds since results are updated when the search finishes (previously they were only updated by the poll timer, which is why we had to set it so low).