Remove documentDisplayName property
The only purpose of observing this property seems to be to update the window title, but since display name is derived from path, which we also observe, it seems redundant.

Though technically the display name could change because extension is hidden or similar, but this is a bit of an edge-case, and long-term we want file browser and tab bar to observe the document’s displayName property directly.