Workaround for window order when restoring session and opening document
While debugging the problem I dumped NSApp.orderedWindows, which turned out to fix the issue. Wouldn’t rely too much on this, but I am not sure what a proper workaround should be, given that this appears to be a framework issue.

For issue description see comment in the source code.