Don’t let untargeted global settings eclipse targeted default settings
For example we have targeted defaults to set tab settings for Ruby, Python, and Makefiles. If the user changed the tab size, it would result in a global untargeted setting that would then eclipse the targeted settings because these were “less local”.

This was not an issue prior to 431bba9a2eb5afb08091c266b5f42b23512c1863.