Limit input to CFCharacterSetIsLongCharacterMember
This is an optimization but it also fixes a crash when CFCharacterSetIsLongCharacterMember is called with extremely large values.

One crash report shows the input being 0x1001DEBC, which is not valid UTF-32, could perhaps be the result of loading a garbage file, so it might make sense to perform some range checks when the user selects to load a file using a UTF encoding.

It could also be command output or possibly copy/paste.