Fix bug in CommonAncestor function
When all paths had the same prefix as first item, the first item would be returned as the common ancestor even if the other paths were not children.

For example:

    /path/to/some_folder
    /path/to/some_folder_with_same_prefix

Here `/path/to/some_folder` was returned even though `/path/to` is the common ancestor.

The function is still not correct when all paths are actual descendants of first path, but since it’s only used for display purposes, and with “find in folder” (where paths shouldn’t overlap), I am not bothering with that issue.