Disable exact rank calculation for large strings
The algorithm to calculate exact rank requires n × m storage which is stack allocated, so for large strings, we could blow the stack.

Rather than switch to dynamic allocation, we’re just foregoing the exact rank, since for large strings, it’s unlikely to be useful to try to calculate such a thing.