Only use dirname_r and basename_r if on macOS 10.12
Not only do we test for the presence of these functions at runtime, but we do not use them at all, when building with the 10.11 SDK (or earlier).

This is because if we had to support building with an earlier SDK it is not enough to declare the prototypes, we also need to add the symbols as weak to the linker options.

Not using the functions does not remove crucial functionality.