Revert "Add ‘public.data’ to list of supported file types for QuickLook plug-in"
This causes TextMate’s QL generator to take over thumbnail generation of all binary files (without a dedicated QuickLook plug-in), for example disk image files.

There does not seem to be any way to dynamically opt out of thumbnail generation, it’s all or nothing, therefor we would either need to completely remove support for thumbnail generation or not support QuickLook previews for files without extension.

For now it’s the latter.

This reverts commit fd827fb8179cd74e553680d605d2172e05822e07.