Update FSItem’s finderTags property when updating the tags on disk
The file browser is observing the `finderTags` property so we need to do KVC updating of this property when we change the tags on disk.

It worked before because changing tags on disk triggers a reload, but that may not always be the case, e.g. the SCM data source showing uncommitted changes did not update in response to changing tags.