Add trailing separator to file browser’s action menu
When the menu is brought up as a context-sensitive menu (right-clicking a file) the system will add a “Services” submenu as the last item, which looks nicer when there is a separator before it.

This submenu is not added when the menu is brought up via the gear button in the status bar, but the system will automatically strip trailing separator items from menus, so in this case, the separator will not be shown at all.