Provide OFBFinderTagsChooser with non-favorite tags
Since OFBFinderTagsChooser always show only the favorite tags, any non-favorite tag set as selected or removable will simply be ignored, so there is no reason to filter these out at the call-site.

In fact, whether or not to show non-favorite tags should probably be an OFBFinderTagsChooser decision.

Furthermore, the tag’s favorite status is set during initialization, so it could actually be outdated when showing the OFBFinderTagsChooser.