Change definition of moving to begin/end of highlight characters
Previously if caret was before an “open” character or after a “close” character, example:

    if‸(true)‸

Then we would move to the matching character, rather than the begin/end character of the “parent” block.

We no longer move to the matching character but instead go straight to the begin/end of parent block.