Set a foreground color when styling is disabled (find text fields)
We remove all styles from the string, even when styling is disabled, as these styles could be from when styling was enabled, so we can’t trust them to be proper defaults.

However, a string with no styles will be rendered as black in the interface, rather than using the default text color for the interface, therefore we now explicitly set the string’s foreground color to be the control text color.