Add helper function to create document icon images
This is based on the somewhat new block-based NSImage rendering API which simplifies things tremendously compared to our existing OakFileIconImage.

Like with the SCMManager class, this new API is limited to a future file browser, as for example images need to update themselves when SCM status or document’s modified state changes, but the NSImage class itself has no way to indicate that it needs to be redrawn, therefore the places we show images should bind to a file item which will then be responsible for updating its image property when the document state changes.