Add FileItem class as interface to the filesystem
FileItem objects will automatically update their SCMStatus and icon based on the properties of the object, but the item’s children (and arranged children) needs to be set externally.

The reason is that to properly sync insert/remove animations with changes to a item’s children, updating the children should be done by the same controller that updates the outline view.