Don’t move focus to text view by default when opening/closing documents
For example when implicitly opening documents because we are duplicating the active document in the file browser, or closing a document because its file got deleted in the file browser, we want focus to stay in the file browser.

Previously we refocused the file browser when closing documents as a result of deleting files.

For duplicating files, this seemed to work because the file browser would not start editing the new file name until it had detected disk changes, thus it would happen after a slight delay.