Disable animated file browser updates when we have more than ~500 items
Insert and remove could still be done, but the re-ordering does not scale as we currently use an LCS algorithm that require N*M storage.

Long-term my thinking is that we can limit animated re-ordering to only include the visible rows, but my initial attempt at this showed some issues with the NSOutlineView, as re-ordering non-visible items still require a reload, which had visible side-effects.