Set titlebarAppearsTransparent for main window
This makes the title bar render in the same color as it did prior to building on 10.14, so that the current tab colors match.

It does however make sheets appear from “above” the title bar and makes the Dark Mode style darker, so this is just an intermediate fix until the tab bar gets revamped for 10.14.