Let text/background color depend on effective appearance in about window
I was unable to update the CSS from Objective-C (using evaluateJavaScript:) which is why the content has to request the styles from the exposed TextMate object.

It would be best if this was implicit (which I know is easy with WKWebView, so better just migrate to WKWebView than fight more with WebView).

Though it is probably better to have separate style sheets for light and dark, eventually we may even be able to rely on media queries for the appearance.