History boxes retain their selection consistently
Both the find and replace boxes now have identical behavior when something in their list is selected: it remains selected in the list, but when the list is opened again it is automatically scrolled to the top. This is under the assumption that the user is probably most interested in whatever they've searched for most recently, rather than whatever happens to be just below their current selection in the history list.