Do not set button’s hugging priority in OakCreateButton
We generally do want buttons to hug their content, but as long as there is flexible space in the layout, they should act as desired, as the flexible will get the superfluousness width/height.

I checked several dialogs and found none that breaks with this change, presumably it was required for some layout that no longer exists.