Use NSButton/NSTextField convenience constructors on 10.12
The convenience constructors initializes several additional properties and should be future compatible.

It’s not ideal to have two code paths to construct the UI, but these helper functions should have stabilized by now and will soon be dropped, so this commit also serves to ensure that our UI works with the new convenience constructors.