Do not call SCM status callbacks when status changes from unknown → none
We ship with the “Update Gutter” command enabled by default which is causing a small delay when opening documents because bundle items are currently blocking the main thread.

This commit (together with the previous one) makes it so that we only call it when we have an assumption that it will actually update the gutter.