Add one line of margin below tab views
At least on macOS 10.14 the auxiliary titlebar view is placed so that the bottom line of the view overlaps with the divider that separates the titlebar from the window content.

While the tab view background color is using alpha, the divider was still visible, but it didn’t look that good when using the light appearance.