Avoid ‘+=’ with “local” variables in build file
By eliminating this operator we can treat the build files as plain dictionaries (no merging of re-declared values). While we still use the append operator in most of the files, it is appending to externally defined variables, so this can be changed to simple assignment, as long as the tree of build files is available to the processor.