Use __DATE__ instead of (custom) COMPILE_DATE macro
I prefer the YYYY-MM-DD format of our custom macro, but it has a problem with precompiled headers in that the macro changes value daily, and so, if specified when precompiling headers, they become invalid the next day.

Previously we solved it by declaring the macro only for application targets, but the new build system is not hierarchical in the same way and does not currently support this.