The $1-n variables in nested replacements weren’t cleared properly
When searching for a regular expression with captures, we have $1-n available in the replacement (format) string.

If this format string does further replacements with optional captures (e.g. /^(_)?foo/) then $1 would be inherited from the parent match, when the optional capture did not match anything.

This bug was introduced in 0b1496011ffd1accd36c74843b485e47b375be98.