Use NSFileManager’s replaceItemAtURL:… as default way to save atomically
NSFileManager’s approach is inferior to exchangedata(), but as this feature is no longer supported with Apple’s default file system (APFS), there is little reason to maintain our own implementation of the fallback code required when we cannot use exchangedata(), especially as the OS know more about file references (aliases), metadata, and other stuff that should be preserved when writing a new (replacement) file.

In theory the NSFileManager strategy can fail to create the temporary destination, in which case we will pass NULL to open() and crash.

Will fix this in a later commit.