Add OakTabBarViewController to simplify using OakTabBarView
This is a subclass of NSTitlebarAccessoryViewController so it can be added directly to windows as an accessory view controller.

It is key/value bindings-compliant for its properties, which should be bound directly to an NSArrayController, which will make the tab bar view reflect the objects managed by this array controller, including updating the tab bar view when object properties such as title or modified state changes.